Move default_create_cli_opts fixture to conftest.py and deduplicate
Move the shared CreateCliOptions fixture from create_test.py to
conftest.py so both test files can use it. Update the test in
test_create.py to use model_copy with to_update_dict to override
only the 6 fields that differ, replacing the 88-line inline
construction.
